requirements.txt
openai
chromadb
python-docx
python-dotenv
streamlit
